python-flask/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3




blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
Flask-Bcrypt==1.0.1  # For bcrypt password hashing (optional)
Flask-JWT-Extended==4.4.0  # For managing JWT tokens securely
Flask-Limiter==2.5.1  # For rate limiting requests (optional)
python-dotenv==1.0.0  # For loading environment variables from .env (optional)
pytest==7.3.1
pytest-cov==4.0.0
Flask-Testing==0.8.1

